requirements.txt
pyzmq==25.1.2
psutil==5.9.8
influxdb-client==1.41.0 